// HelpDeskFE/HelpDeskFETp/src/api/model/createAlertConfigurationDto.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * HelpDesk API
 * OpenAPI spec version: v1
 */

export interface CreateAlertConfigurationDto {
  /**
     * @minLength 1
     * @maxLength 150
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  subject: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  description: string;
  isGlobal?: boolean;
  isActive?: boolean;
  /** @nullable */
  idArea?: number | null;
  /** @nullable */
  idAgency?: number | null;
  /** @nullable */
  scheduledDate?: string | null;
}